refactor : 게시글 생성 전략 패턴 적용

+++ storage/db-core/src/main/java/com/cos/cercat/database/common/util/PagingUtil.java
@@ -13,7 +13,7 @@ import java.time.LocalDateTime;
 
 public class PagingUtil {
 
-    public static OrderSpecifier<?> getOrderSpecifier(Pageable pageable, DateTimePath<LocalDateTime> createdAt, NumberPath<Integer> likeCount) {
+    public static OrderSpecifier<?> getOrderSpecifier(Pageable pageable, DateTimePath<LocalDateTime> createdAt, NumberPath<Long> likeCount) {
         if (!pageable.getSort().isEmpty()) {
             for (Sort.Order order : pageable.getSort()) {
                 Order direction = order.getDirection().isAscending() ? Order.ASC : Order.DESC;
